Deepfake-Video-Detector-Autoencoder/requirements.txt
torch>=1.9.0
torchvision>=0.11.0
opencv-python>=4.5.3
facenet-pytorch>=2.5.2
numpy>=1.19.5
streamlit>=1.0.0
PyYAML>=5.4.1
tqdm>=4.62.0
matplotlib>=3.4.3
Pillow>=8.3.1
efficientnet_pytorch>=0.7.1
